backend/requirements.txt
# Core Framework
Flask==3.0.0
flask-cors==4.0.0
Flask-Caching==2.1.0

# Database
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0

# API & HTTP
requests==2.31.0

# Scheduler (for auto-import)
schedule==1.2.0
